i think this huge refactor is finished now

+++ Assets/framework/source/System/Resources.cs
@@ -24,22 +24,22 @@ namespace Framework.Data {
 			[Serializable]
 			public struct Dependency {
 				[JsonProperty] public string Type;
-				[JsonProperty] public string ID;
+				[JsonProperty] public string UID;
 			}
 
-			[JsonIgnore] public string ID => id;
+			[JsonIgnore] public string UID => uid;
 			[JsonIgnore] public string Name => name;
 			[JsonIgnore] public string Path => path;
 			[JsonIgnore] public Dependency[] Dependencies => dependencies;
 
-			public void SetResourceData ( string id, string name, string path ) {
+			public void SetResourceData ( string uid, string name, string path ) {
 
-				this.id = id;
+				this.uid = uid;
 				this.name = name;
 				this.path = path;
 			}
 
-			[JsonProperty] private string id;
+			[JsonProperty] private string uid;
 			[JsonProperty] private string name;
 			[JsonProperty] private string path;
 			[JsonProperty] private Dependency[] dependencies;
@@ -47,7 +47,7 @@ namespace Framework.Data {
 
 			// ********** IEquatable Implementation **********
 
-			bool IEquatable<Resource>.Equals ( Resource other ) => other.ID.Equals( this.ID );
+			bool IEquatable<Resource>.Equals ( Resource other ) => other.UID.Equals( this.UID );
 		}
 	}
 
@@ -69,25 +69,25 @@ namespace Framework.Data {
 
 		public event Framework.Event<IList<TMetadata>>.Signature OnMetadataChanged;
 
-		public Resources ( string resourcePath, string resourceExtension, int idLength ) {
+		public Resources ( string resourcePath, string resourceExtension, int uidLength ) {
 
 			// init data structures
 			_allMetadata = new List<TMetadata>();
 			_allResourceNames = new HashSet<string>();
-			_metadataById = new Dictionary<string, TMetadata>();
-			_loadedResourcesById = new Dictionary<string, TResource>();
+			_metadataByUID = new Dictionary<string, TMetadata>();
+			_loadedResourcesByUID = new Dictionary<string, TResource>();
 
 			// store ivars
 			_path = resourcePath;
 			_extension = resourceExtension;
-			_idLength = idLength;
+			_uidLength = uidLength;
 		}
 		public void Close () {
 
 			SaveMetadataToDisk();
-			var loadedResourceIDs = new List<string>( _loadedResourcesById.Keys );
-			foreach ( var id in loadedResourceIDs ) {
-				UnloadResource( id, save: true );
+			var loadedResourceUIDs = new List<string>( _loadedResourcesByUID.Keys );
+			foreach ( var uid in loadedResourceUIDs ) {
+				UnloadResource( uid, save: true );
 			}
 		}
 
@@ -114,31 +114,33 @@ namespace Framework.Data {
 			var metadata = new TMetadata();
 			var resource = new TResource();
 			var resourceID = StringHelpers.UID.Generate(
-				length: _idLength,
-				validateUniqueness: id => _metadataById.KeyIsUnique( id )
+				length: _uidLength,
+				validateUniqueness: uid => _metadataByUID.KeyIsUnique( uid )
 			);
 
 			// init metadata
 			metadata.SetResourceData(
-				id: resourceID,
+				uid: resourceID,
 				name: name,
 				path: ResourcePath( name )
 			);
 
 			// store
 			TrackMetadata( metadata );
-			_loadedResourcesById[resourceID] = resource;
+			_loadedResourcesByUID[resourceID] = resource;
 
 			return (metadata, resource);
 		}
-		public bool Delete ( string id ) {
+		public bool Delete ( string uid ) {
+
+			if ( uid == null ) { return false; }
 
 			try {
 
-				var metadata = RequestMetadata( id );
+				var metadata = RequestMetadata( uid );
 
 				UntrackMetadata( metadata );
-				UnloadResource( id, save: false );
+				UnloadResource( uid, save: false );
 
 				PersistentStore.Delete( metadata.Path );
 
@@ -146,93 +148,93 @@ namespace Framework.Data {
 
 			} catch ( ResourceMetadataNotFoundException ) {
 
-				Debug.LogError( $"Client.Model.Resources.Delete: Can't delete, metadata for id {id} doesn't exist." );
+				Debug.LogError( $"Client.Model.Resources.Delete: Can't delete, metadata for uid {uid} doesn't exist." );
 				return false;
 			}
 		}
 
 		// resource loading
-		public bool LoadResource ( string id ) {
+		public bool LoadResource ( string uid ) {
 
 			// early exit
-			if ( ResourceIsLoaded( id ) ) { return true; }
+			if ( ResourceIsLoaded( uid ) ) { return true; }
 
 			try {
 
-				var metadata = RequestMetadata( id );
+				var metadata = RequestMetadata( uid );
 				var resource = PersistentStore.Load<TResource>( metadata.Path );
-				_loadedResourcesById[id] = resource;
+				_loadedResourcesByUID[uid] = resource;
 				return true;
 
 			} catch ( ResourceMetadataNotFoundException ) {
 
-				Debug.LogError( $"Client.Model.Resources.Load: Couldn't find resource metadata for id {id}" );
+				Debug.LogError( $"Client.Model.Resources.Load: Couldn't find resource metadata for uid {uid}" );
 				return false;
 			}
 			// TODO: handle deserialization exceptions
 		}
-		public void UnloadResource ( string id, bool save = true ) {
+		public void UnloadResource ( string uid, bool save = true ) {
 
-			if ( !ResourceIsLoaded( id ) ) {
+			if ( !ResourceIsLoaded( uid ) ) {
 				return;
 			}
 
 			if ( save ) {
-				SaveResource( id );
+				SaveResource( uid );
 			}
 
-			_loadedResourcesById.Remove( id );
+			_loadedResourcesByUID.Remove( uid );
 		}
-		public bool SaveResource ( string id ) {
+		public bool SaveResource ( string uid ) {
 
 			try {
 
-				var metadata = RequestMetadata( id );
-				var resource = RequestResource( id, load: false );
+				var metadata = RequestMetadata( uid );
+				var resource = RequestResource( uid, load: false );
 				PersistentStore.Save<TResource>( metadata.Path, resource );
 				return true;
 
 			} catch ( ResourceNotLoadedException ) {
-				Debug.LogError( $"Client.Model.Resources.Save: Couldn't find loaded resource for id {id}" );
+				Debug.LogError( $"Client.Model.Resources.Save: Couldn't find loaded resource for uid {uid}" );
 				return false;
 
 			} catch ( ResourceMetadataNotFoundException ) {
-				Debug.LogError( $"Client.Model.Resources.Save: Couldn't find resource metadata for id {id}" );
+				Debug.LogError( $"Client.Model.Resources.Save: Couldn't find resource metadata for uid {uid}" );
 				return false;
 			}
 		}
 
 		// data requests
-		public TMetadata RequestMetadata ( string id ) {
+		public TMetadata RequestMetadata ( string uid ) {
 
-			if ( _metadataById.TryGetValue( id, out var metadata ) ) {
+			if ( _metadataByUID.TryGetValue( uid, out var metadata ) ) {
 				return metadata;
 			} else {
 				throw new ResourceMetadataNotFoundException();
 			}
 		}
-		public TResource RequestResource ( string id, bool load ) {
+		public TResource RequestResource ( string uid, bool load ) {
 
 			if ( load ) {
-				LoadResource( id );
+				LoadResource( uid );
 			}
 
 			// if not loaded, determine issue
-			if ( !ResourceIsLoaded( id ) ) {
-				if ( !_metadataById.ContainsKey( id ) ) {
+			if ( !ResourceIsLoaded( uid ) ) {
+				if ( !_metadataByUID.ContainsKey( uid ) ) {
 					throw new ResourceMetadataNotFoundException();
 				} else {
 					throw new ResourceNotLoadedException();
 				}
 			}
 
-			return _loadedResourcesById[id];
+			return _loadedResourcesByUID[uid];
 		}
 
 		// helpers
 		public bool NameIsUnique ( string name ) => !_allResourceNames.Contains( name );
 		public IList<TMetadata> GetAllMetadata () => _allMetadata.AsReadOnly();
-		public List<TResource> GetLoadedResources () => new List<TResource>( _loadedResourcesById.Values );
+		public List<TResource> GetLoadedResources () => new List<TResource>( _loadedResourcesByUID.Values );
 
 
 		// *********** Private Interface ***********
@@ -242,25 +244,25 @@ namespace Framework.Data {
 
 		// runtime data
 		private HashSet<string> _allResourceNames;
-		private Dictionary<string, TMetadata> _metadataById;
-		private Dictionary<string, TResource> _loadedResourcesById;
+		private Dictionary<string, TMetadata> _metadataByUID;
+		private Dictionary<string, TResource> _loadedResourcesByUID;
 
 		// instance variables
 		private string _path;
 		private string _extension;
-		private int _idLength;
+		private int _uidLength;
 
 		// private helpers
-		private string MetadataPath => $"{_path}/.metadata";
+		private string MetadataPath => $"{_path}/{_extension}.metadata";
 		private string ResourcePath ( string name ) => $"{_path}/{name}.{_extension}";
-		private bool ResourceIsLoaded ( string id ) => _loadedResourcesById.ContainsKey( id );
+		private bool ResourceIsLoaded ( string uid ) => _loadedResourcesByUID.ContainsKey( uid );
 
 		private void TrackMetadata ( TMetadata metadata ) {
 
 			// track metadata
 			_allMetadata.Add( metadata );
 			_allResourceNames.Add( metadata.Name );
-			_metadataById.Add( metadata.ID, metadata );
+			_metadataByUID.Add( metadata.UID, metadata );
 
 			// fire event
 			Framework.Event<IList<TMetadata>>.Fire(
@@ -275,7 +277,7 @@ namespace Framework.Data {
 			// untrack metadata
 			_allMetadata.Remove( metadata );
 			_allResourceNames.Remove( metadata.Name );
-			_metadataById.Remove( metadata.ID );
+			_metadataByUID.Remove( metadata.UID );
 
 			// fire event
 			Framework.Event<IList<TMetadata>>.Fire(
@@ -292,12 +294,12 @@ namespace Framework.Data {
 
 			// clear runtime data
 			_allResourceNames?.Clear();
-			_metadataById?.Clear();
+			_metadataByUID?.Clear();
 
 			// copy serialized data into runtime structures
 			_allMetadata.ForEach( metadata => {
 				_allResourceNames.Add( metadata.Name );
-				_metadataById.Add( metadata.ID, metadata );
+				_metadataByUID.Add( metadata.UID, metadata );
 			} );
 		}
 	}
